services/llm_agent/requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.6
paho-mqtt==2.1.0
pydantic==2.8.2
httpx==0.27.0
openai==1.12.0
anthropic==0.8.1
python-dotenv==1.0.0
